Fix handling of functions with multiple definitions
FunctionDeclMatchHandler::finalize(), working at a single TU at a time,
accumulates dependencies in two separate 'Uses' 'Defs' sets:

* Uses: All functions that are used but not defined in the TU
* Defs: All functions that are defined but not used in the TU

This lacks the case when a function is defined _and_ used in the TU
at the same time. Such function was omitted and could later be
considered as 'unused', e.g., if another definition was found
in a tests 'stub' file.