Fix handling of functions with definition-only decorations (#20)
* Fix handling of functions with const parameter variables

Apart from resolving this specific problem, reworked and fixed
FunctionDeclMatchHandler::finalize() that analyzes dependencies.
While working at a single TU at a time, it accumulated the
dependencies in the two separate 'Uses' 'Defs' sets:

* Uses: All functions that are used but not defined in the TU
* Defs: All functions that are defined but not used in the TU

This lacks the case when a function is defined and used in the TU
at the same time. Such function was omitted and could later be
considered as 'unused', e.g., if another definition was found
in a 'stub' file.

* Undone out-of-scope changes

that should be placed in a separate PR.

* Keep using F->getDefinition() in finalize() for USR unrelated code

* Added a test

* Polished