fixup: Do not redirect stderr
LLVM's `lit` program collects and reports RUN's stdout _and_ stderr.
Most tests do redirect stderr, but they do it to check stderr contents
with FileCheck. This test case does not use FileCheck because this test
is about exit code rather than accompanying output.

We could use FileCheck and a CHECK rule to [more-or-less] ensure that
the compilation fails specifically due to our `#error` preprocessor
directive. In that case, we would have to redirect, but we KISS for now.